pybullet_sim_realman_RM65/requirements.txt
# 核心依赖
pybullet>=3.2.5
numpy>=1.21.0
scipy>=1.7.0
torch>=1.9.0
torchvision>=0.10.0
gymnasium>=0.28.0

# 机器学习和深度学习
scikit-learn>=1.0.0
opencv-python>=4.5.0
Pillow>=8.3.0

# 可视化和绘图
matplotlib>=3.4.0
seaborn>=0.11.0

# 数据处理
pandas>=1.3.0

# 日志和监控
tensorboard>=2.7.0
tqdm>=4.62.0

# 配置和工具
pyyaml>=5.4.0
jsonschema>=3.2.0

# 可选依赖（用于高级功能）
# jupyter>=1.0.0  # 用于交互式开发
# plotly>=5.0.0   # 用于交互式可视化
# wandb>=0.12.0   # 用于实验跟踪